+++ CHANGES.rst
@@ -2,7 +2,7 @@
 Revision 0.2.1, XX-06-2017
 --------------------------
 
-- ZIP archive reader implemented
+* ZIP archive reader implemented
 
 Revision 0.1.3, 19-05-2017
 --------------------------